+++ inst/NEWS.md
@@ -10,3 +10,13 @@
 
 * We have added explicit dependencies to src/Makevars 
 	to tell make the constraints on the order of compilation.	
+
+### 0.1.2 (2022-11-16)
+
+* Minor fix
+
+* 00_lb.f90 was merged with 00_lbfgsb.f90
+
+* We have added more information to src/Makevars in order to 
+	setup dependencies for parallel make.	
+